backend/requirements-prod.txt
fastapi
uvicorn[standard]
gunicorn
sqlmodel
pydantic
python-multipart
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.2
pymysql
pytest
httpx
python-dotenv
redis
requests
openai
reportlab
captcha
pycryptodome
